Patch Note v2.1.0
**Smash Or Pass** :
- Pour des questions d'optimisation et de temps de chargement des menus, les images seront automatiquement transformée en webp et à 1Mo.
:warn: Cela n'affecte pas le mode preview (display de plusieurs tenues en même temps) qui lui créer une copie de l'image 1024x1024
Cependant les gifs ne sont désormais plus autorisé et seront transformé en image fixe.

- Mettre plusieurs fichiers lors du changement d'une image d'outfit ne fera plus une erreur :thumbsup:

- Je ne l'avais pas préciser mais lors du hotfix post v2.0.1 j'ai ajouter un filet de sécurité pour éviter de crash lors d'une majeure.
Cependant les erreurs fatales font toujours reboot le bot pour éviter des boucles infinies de crash qui pourrais paralysé le bot.

- Les menus ne sont désormais plus infini et se fermeront automatiquement au bout de 10 minutes d'inactivité.
Laisser un tel menu ouvert est un véritable gouffre en ressources et beaucoup oublient de le fermer :smiling_face_with_tear:

- En réponse à ce changement : Le menu de fin s'est vue apporter un changement
  - Le bouton "Terminer" s'appelle désormais "Publier" et permet de publier la même interface de fin qu'avant sans devoir ré-ouvrir un menu.
  - Cette interface de fin est désormais interactive et vous pourrez visualiser les résultats sans devoir télécharger le fichier, pratique non ?

- Ajout du mode téléphone
  C'est une petite icone de téléphone à On/Off sur le menu principale, ça permet de compacter un peu plus certains texte et ne display les listes que sur une seule colonne

- Le musée des personnages ouvre ses portes :fire:
  Cette page vous permet une page qui permet de voir le score et le ratio des personnages ainsi que leurs tenus!

Patch Note v2.1.1
**Smash Or Pass** :
- En mode gestion des personnages : La liste des personnages est désormais triée a-z

+++ events/buttons.js
@@ -1,12 +1,13 @@
 import { AttachmentBuilder, Events, MessageFlags } from 'discord.js';
 import { Registry } from '#modules/Registry';
 import { dbManager } from '#modules/database/Manager';
-import { noop, IsMessageAuthorAdmin, isDefined, isObject, MD5 } from '#modules/Utils';
+import { noop, IsMessageAuthorAdmin, isDefined, isObject, MD5, generateEasyPassword, ModalForm } from '#modules/Utils';
 import util from "util"
 import Locale from '#modules/Locales';
 import { PERMISSION } from '#constants';
 import path from 'path';
 import fs from 'fs';
+import { GameSmashOrPassFinalViewer } from '#modules/menus/sop/index';
 
 Registry.register({
   name: "Event: InteractionCreate (Buttons)",
@@ -39,7 +40,7 @@ export default {
   name: Events.InteractionCreate,
   run: async ({ client, parameters: [interaction] }) => {
     if (!interaction.isButton()) return;
-    if (!/(GETLOG|GETFILE|DELETE):[a-z-0-9]+(:[a-z-0-9]+)*/gmi.test(interaction.customId)) return;
+    if (!/(GETLOG|GETFILE|DELETE|MENU-SOP-FINAL):[a-z-0-9]+(:[a-z-0-9]+)*/gmi.test(interaction.customId)) return;
 
     let userPermission = PERMISSION.USER;
     if (interaction.channel.type !== 'DM' && IsMessageAuthorAdmin(interaction, false)) userPermission = PERMISSION.GUILD_ADMIN;
@@ -114,7 +115,22 @@ export default {
           break;
         }
         case "DELETE": {
+          const password = Array.from(Array(3), () => generateEasyPassword(2)).map(s => s.ucFirst()).join('');
+
+          let modal = new ModalForm({ title: `Veuillez confirmez en tapant le mot de passe`, time: 60_000 })
+            .addRow().addTextField({ name: 'password', label: `Veuillez taper ${password} pour confirmer`, placeholder: password })
+          ;
+          
+          let result = await modal.setInteraction(interaction).popup();
+          if (!result || !result.get('password')) return;
+          if (result.get('password') !== password) return;
+
           interaction.message?.delete().catch(noop);
+          break;
+        }
+        case "MENU-SOP-FINAL": {
+          // interaction.message?.delete().catch(noop);
+          GameSmashOrPassFinalViewer({ client, interaction });
           break;
         }
       }
